fix(FEC-13628): markers not shown on timeline for clipped video (#32)
### Description of the Changes

- **bugfix**

**the issue:**
sometimes markers/chapters are not shown on timeline.

**root cause:**
when checking if the duration is correct, it is possible that `seekFrom` has value of `0`.
currently, when the value is 0, the logic breaks and the new calculation for duration is not happening. As a result, the duration is never correct and `timelineDurationPromise` is never resolved. This affect rendering cuepoints on timeline.

**solution:**
validate `typeof seekFrom`.

Solves [FEC-13628](https://kaltura.atlassian.net/browse/FEC-13628)